requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.13.3
numpy==2.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.40.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==5.0.3
